airbus/requirements.txt
requests
boto3
kubernetes
pulsar-client
pycryptodome